Refactor to expose forward declaration in header.
