+++ include/nanospline/forward_declaration.h
