+++ include/nanospline/forward_declaration_inflection.h
