+++ include/nanospline/forward_declaration_singularity.h
