+++ scripts/inflection.py
@@ -205,7 +205,7 @@ extern_definition_template = """/**
 #include <cassert>
 #include <vector>
 
-#include "forward_declaration_inflection.h"
+#include <nanospline/forward_declaration_inflection.h>
 
 namespace nanospline {{
 namespace internal {{
@@ -373,12 +373,12 @@ if __name__ == "__main__":
             f.write(code)
 
     body = extern_declaration_template.format(body=("\n".join(extern_declarations)))
-    with open(os.path.join(dir_path, "..", "tests",
+    with open(os.path.join(dir_path, "..", "include", "nanospline",
         "forward_declaration_inflection.h"), "w") as f:
         f.write(body)
 
     body = extern_definition_template.format(body=("\n".join(extern_definitions)))
-    with open(os.path.join(dir_path, "..", "tests",
+    with open(os.path.join(dir_path, "..", "src",
         "forward_declaration_inflection.cpp"), "w") as f:
         f.write(body)
