+++ src/forward_declaration_inflection.cpp
@@ -1,11 +1,11 @@
 /**
- * This code is automatically generated by scripts/optimal_turning_angle.py
+ * This code is automatically generated by scripts/inflection.py
  */
 
 #include <cassert>
 #include <vector>
 
-#include "forward_declaration_match_tangent.h"
+#include <nanospline/forward_declaration_inflection.h>
 
 namespace nanospline {
 namespace internal {
@@ -14,20 +14,18 @@ namespace internal {
 #if defined(HIGH_DEGREE_SUPPORT) || 2 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_RationalBezier_degree_2(
+std::vector<Scalar> compute_RationalBezier_degree_2_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2,
         Scalar w0, Scalar w1, Scalar w2,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
 #define Scalar float
-template
-std::vector<Scalar> match_tangent_RationalBezier_degree_2(
+extern template
+std::vector<Scalar> compute_RationalBezier_degree_2_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2,
         Scalar w0, Scalar w1, Scalar w2,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
@@ -37,20 +35,18 @@ std::vector<Scalar> match_tangent_RationalBezier_degree_2(
 #if defined(HIGH_DEGREE_SUPPORT) || 3 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_RationalBezier_degree_3(
+std::vector<Scalar> compute_RationalBezier_degree_3_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3,
         Scalar w0, Scalar w1, Scalar w2, Scalar w3,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
 #define Scalar float
-template
-std::vector<Scalar> match_tangent_RationalBezier_degree_3(
+extern template
+std::vector<Scalar> compute_RationalBezier_degree_3_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3,
         Scalar w0, Scalar w1, Scalar w2, Scalar w3,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
@@ -60,20 +56,18 @@ std::vector<Scalar> match_tangent_RationalBezier_degree_3(
 #if defined(HIGH_DEGREE_SUPPORT) || 4 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_RationalBezier_degree_4(
+std::vector<Scalar> compute_RationalBezier_degree_4_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4,
         Scalar w0, Scalar w1, Scalar w2, Scalar w3, Scalar w4,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
 #define Scalar float
-template
-std::vector<Scalar> match_tangent_RationalBezier_degree_4(
+extern template
+std::vector<Scalar> compute_RationalBezier_degree_4_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4,
         Scalar w0, Scalar w1, Scalar w2, Scalar w3, Scalar w4,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
@@ -83,189 +77,189 @@ std::vector<Scalar> match_tangent_RationalBezier_degree_4(
 #if defined(HIGH_DEGREE_SUPPORT) || 2 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_2(
+std::vector<Scalar> compute_Bezier_degree_2_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 2 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_2(
+std::vector<Scalar> compute_Bezier_degree_2_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
 
 
 #if defined(HIGH_DEGREE_SUPPORT) || 3 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_3(
+std::vector<Scalar> compute_Bezier_degree_3_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 3 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_3(
+std::vector<Scalar> compute_Bezier_degree_3_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
 
 
 #if defined(HIGH_DEGREE_SUPPORT) || 4 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_4(
+std::vector<Scalar> compute_Bezier_degree_4_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 4 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_4(
+std::vector<Scalar> compute_Bezier_degree_4_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
 
 
 #if defined(HIGH_DEGREE_SUPPORT) || 5 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_5(
+std::vector<Scalar> compute_Bezier_degree_5_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 5 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_5(
+std::vector<Scalar> compute_Bezier_degree_5_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
 
 
 #if defined(HIGH_DEGREE_SUPPORT) || 6 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_6(
+std::vector<Scalar> compute_Bezier_degree_6_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 6 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_6(
+std::vector<Scalar> compute_Bezier_degree_6_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
 
 
 #if defined(HIGH_DEGREE_SUPPORT) || 7 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_7(
+std::vector<Scalar> compute_Bezier_degree_7_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6, Scalar cx7, Scalar cy7,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 7 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_7(
+std::vector<Scalar> compute_Bezier_degree_7_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6, Scalar cx7, Scalar cy7,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
 
 
 #if defined(HIGH_DEGREE_SUPPORT) || 8 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_8(
+std::vector<Scalar> compute_Bezier_degree_8_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6, Scalar cx7, Scalar cy7, Scalar cx8, Scalar cy8,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 8 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_8(
+std::vector<Scalar> compute_Bezier_degree_8_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6, Scalar cx7, Scalar cy7, Scalar cx8, Scalar cy8,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
 
 
 #if defined(HIGH_DEGREE_SUPPORT) || 9 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_9(
+std::vector<Scalar> compute_Bezier_degree_9_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6, Scalar cx7, Scalar cy7, Scalar cx8, Scalar cy8, Scalar cx9, Scalar cy9,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 9 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_9(
+std::vector<Scalar> compute_Bezier_degree_9_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6, Scalar cx7, Scalar cy7, Scalar cx8, Scalar cy8, Scalar cx9, Scalar cy9,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
 
 
 #if defined(HIGH_DEGREE_SUPPORT) || 10 < 5
 #define Scalar double
 template
-std::vector<Scalar> match_tangent_Bezier_degree_10(
+std::vector<Scalar> compute_Bezier_degree_10_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6, Scalar cx7, Scalar cy7, Scalar cx8, Scalar cy8, Scalar cx9, Scalar cy9, Scalar cx10, Scalar cy10,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
 
+#if 10 < 10
 #define Scalar float
 template
-std::vector<Scalar> match_tangent_Bezier_degree_10(
+std::vector<Scalar> compute_Bezier_degree_10_inflections(
         Scalar cx0, Scalar cy0, Scalar cx1, Scalar cy1, Scalar cx2, Scalar cy2, Scalar cx3, Scalar cy3, Scalar cx4, Scalar cy4, Scalar cx5, Scalar cy5, Scalar cx6, Scalar cy6, Scalar cx7, Scalar cy7, Scalar cx8, Scalar cy8, Scalar cx9, Scalar cy9, Scalar cx10, Scalar cy10,
-        const Eigen::Matrix<Scalar, 2, 1>& tangent,
         Scalar t0,
         Scalar t1);
 #undef Scalar
+#endif
 #endif
