+++ tests/test_BSplinePatch.cpp
@@ -3,6 +3,7 @@
 
 #include <nanospline/BSplinePatch.h>
 #include <nanospline/save_obj.h>
+#include <nanospline/forward_declaration.h>
 
 #include "validation_utils.h"
