+++ tests/test_Bezier.cpp
@@ -3,8 +3,8 @@
 
 #include <nanospline/Bezier.h>
 #include <nanospline/save_svg.h>
+#include <nanospline/forward_declaration.h>
 
-#include "forward_declaration.h"
 #include "validation_utils.h"
 
 TEST_CASE("Bezier", "[nonrational][bezier]") {
