+++ tests/test_NURBS.cpp
@@ -4,8 +4,8 @@
 #include <iostream>
 #include <nanospline/NURBS.h>
 #include <nanospline/save_svg.h>
+#include <nanospline/forward_declaration.h>
 
-#include "forward_declaration.h"
 #include "validation_utils.h"
 
 TEST_CASE("NURBS", "[rational][nurbs][bspline]") {
