+++ tests/test_conversion.cpp
@@ -2,7 +2,7 @@
 
 #include <nanospline/conversion.h>
 #include <nanospline/save_svg.h>
-#include "forward_declaration.h"
+#include <nanospline/forward_declaration.h>
 #include "validation_utils.h"
 
 TEST_CASE("conversion", "[conversion]") {
