+++ tests/test_optimal_turing_angle.cpp
@@ -4,8 +4,7 @@
 #include <nanospline/optimal_turning_angle.h>
 #include <nanospline/split.h>
 #include <nanospline/save_svg.h>
-
-#include "forward_declaration.h"
+#include <nanospline/forward_declaration.h>
 
 TEST_CASE("opt_turn_angle", "[opt_turn_angle]") {
     using namespace nanospline;
